+++ src/main/java/eu/tib/controller/OrcidController.java
@@ -1,5 +1,6 @@
 package eu.tib.controller;
 
+import eu.tib.controller.validation.InputValidator;
 import eu.tib.service.ResponseService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiParam;
@@ -8,10 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.StopWatch;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
@@ -27,18 +25,39 @@ public class OrcidController {
     @Autowired
     private ResponseService responseService;
 
-    @PostMapping(value = "/getCurrentEmployees", produces = "application/json")
-    public ResponseEntity<String> getCurrentEmployees(
-            @Valid @Pattern(regexp = "^Q[1-9]\\d*$")
-            @ApiParam("Wikidata id for a research organization starting with Q")
-            @RequestParam String wikidata) {
+    @GetMapping(value = "/personPlusPublications", produces = "application/json")
+    public ResponseEntity<String> getPersonPlusPublications(
+            @Valid @Pattern(regexp = InputValidator.orcid)
+            @ApiParam("Complete Orcid URL consisting of https://orcid.org/ plus id")
+            @RequestParam String orcid) {
 
-        final String id = "sparqlg/orcid/employees";
-        log.info("Incoming Request for " + id + " with wikidata: " + wikidata);
+        final String id = "sparqlg/orcid/person";
+        log.info("Incoming Request for " + id + " with orcid: " + orcid);
         StopWatch stopWatch = new StopWatch(id);
         stopWatch.start(id);
 
-        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("wikidata", wikidata));
+        String orcid_id = orcid.replaceFirst("https://orcid.org/","");
+        ResponseEntity result = responseService.buildResponse(id,
+                Collections.singletonMap("orcid", orcid_id));
+
+        stopWatch.stop();
+        log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");
+        return result;
+    }
+
+
+    @GetMapping(value = "/currentEmployeesPlusPublications", produces = "application/json")
+    public ResponseEntity<String> getCurrentEmployeesPlusPublications(
+            @Valid @Pattern(regexp = InputValidator.ror)
+            @ApiParam("Complete ROR URL consisting of https://ror.org/ plus id")
+            @RequestParam String ror) {
+
+        final String id = "sparqlg/orcid/employees";
+        log.info("Incoming Request for " + id + " with ror: " + ror);
+        StopWatch stopWatch = new StopWatch(id);
+        stopWatch.start(id);
+
+        ResponseEntity result = responseService.buildResponse(id, Collections.singletonMap("ror", ror));
 
         stopWatch.stop();
         log.info(id + " took " + stopWatch.getTotalTimeSeconds() + "s");