+++ src/main/java/eu/tib/controller/validation/InputValidator.java
@@ -1,0 +1,7 @@
+package eu.tib.controller.validation;
+
+public class InputValidator {
+    public static final String ror = "^https://ror.org/\\d{2}[a-z0-9]{5}\\d{2}$";
+    public static final String orcid = "^https://orcid.org/\\d{4}-\\d{4}-\\d{4}-\\d{3}(\\d|X)$";
+    public static final String doi = "^10\\.[0-9]{4}.*";
+}